Bugfix: Removed a warning in the backend ajax module for high traffic sites.

+++ Classes/Collectors/LogfileList.php
@@ -118,7 +118,7 @@ class LogfileList extends AbstractCollector implements BacktraceConstInterface
             // Getting the basic info.
             $fileinfo['name'] = basename($file);
             $fileinfo['size'] = $this->fileSizeConvert((int)filesize($file));
-            $fileinfo['time'] = date("d.m.y H:i:s", filemtime($file));
+            $fileinfo['time'] = date("d.m.y H:i:s", (int)filemtime($file));
             $fileinfo['id'] = str_replace('.Krexx.html', '', $fileinfo['name']);
             $fileinfo['dispatcher'] = $this->getRoute($fileinfo['id']);
             $fileinfo['meta'] = $this->addMetaToFileInfo($file);
